add tenant provisioning to data plane
