+++ backend/ee/danswer/server/tenants/models.py
@@ -1,0 +1,6 @@
+from pydantic import BaseModel
+
+
+class CreateTenantRequest(BaseModel):
+    tenant_id: str
+    initial_admin_email: str
