+++ backend/ee/danswer/server/tenants/provisioning.py
@@ -1,0 +1,63 @@
+import contextlib
+import os
+from types import SimpleNamespace
+
+from sqlalchemy import text
+from sqlalchemy.orm import Session
+from sqlalchemy.schema import CreateSchema
+
+from alembic import command
+from alembic.config import Config
+from danswer.db.engine import build_connection_string
+from danswer.db.engine import get_async_session
+from danswer.db.engine import get_sqlalchemy_engine
+from danswer.utils.logger import setup_logger
+
+logger = setup_logger()
+
+
+def run_alembic_migrations(schema_name: str) -> None:
+    logger.info(f"Starting Alembic migrations for schema: {schema_name}")
+
+    try:
+        current_dir = os.path.dirname(os.path.abspath(__file__))
+        root_dir = os.path.abspath(os.path.join(current_dir, "..", "..", ".."))
+        alembic_ini_path = os.path.join(root_dir, "alembic.ini")
+
+        # Configure Alembic
+        alembic_cfg = Config(alembic_ini_path)
+        alembic_cfg.set_main_option("sqlalchemy.url", build_connection_string())
+
+        # Mimic command-line options by adding 'cmd_opts' to the config
+        alembic_cfg.cmd_opts = SimpleNamespace()
+        alembic_cfg.cmd_opts.x = [f"schema={schema_name}"]
+
+        # Run migrations programmatically
+        command.upgrade(alembic_cfg, "head")
+
+        # Run migrations programmatically
+        logger.info(
+            f"Alembic migrations completed successfully for schema: {schema_name}"
+        )
+
+    except Exception as e:
+        logger.exception(f"Alembic migration failed for schema {schema_name}: {str(e)}")
+        raise
+
+
+def create_tenant_schema(tenant_id: str) -> None:
+    with Session(get_sqlalchemy_engine()) as db_session:
+        with db_session.begin():
+            db_session.execute(CreateSchema(tenant_id))
+
+
+async def check_schema_exists(tenant_id: str) -> bool:
+    get_async_session_context = contextlib.asynccontextmanager(get_async_session)
+    async with get_async_session_context() as session:
+        result = await session.execute(
+            text(
+                "SELECT schema_name FROM information_schema.schemata WHERE schema_name = :schema_name"
+            ),
+            {"schema_name": tenant_id},
+        )
+        return result.scalar() is not None
